Optimize 3 more methods
